shell.py
This is the main file to run. Make sure lexer.py, shell.py, “input_sourcecode.txt” are all in the same folder location otherwise the main file will run an error. Running main on shell.py will create output.docx with a table of Tokens and corresponding Lexeme in list.